+++ RA/vdata.cpp
@@ -563,7 +563,7 @@ void VesselTypeClass::One_Time(void)
 		_makepath(fullname, NULL, NULL, uclass.Graphic_Name(), ".SHP");
 		((void const *&)uclass.ImageData) = MFCD::Retrieve(fullname);
 
-		((int &)uclass.MaxSize) = 26;
+		((uint8_t &)uclass.MaxSize) = 26;
 	}
 }
